Refactor: Move VideoPlayer to components package
The VideoPlayer composable has been moved from `com.feature.mediaDetails.mediaDetailsUi.ui.comon` to `com.feature.mediaDetails.mediaDetailsUi.ui.comon.components`.
Import statements in `TvShowDetailsScreen.kt` and `MovieDetailsScreen.kt` have been updated to reflect this change.
